Sanbox shade positioning tweaks into a module
